avoid duplicate inplaceSortBy code
- abstract over the concrete representation (Addr/Mba) of the array
- partially applied primRead/Write hide the concrete representation
