+++ basement/Basement/Alg/Foreign/Prim.hs
@@ -6,6 +6,7 @@ module Basement.Alg.Foreign.Prim
     , primIndex64
     , primRead
     , primWrite
+    , primReadWrite
     ) where
 
 import           GHC.Types
@@ -33,3 +34,7 @@ primRead = primAddrRead
 primWrite :: (PrimMonad prim, PrimType ty) => Mutable (PrimState prim) -> Offset ty -> ty -> prim ()
 primWrite = primAddrWrite
 {-# INLINE primWrite #-}
+
+primReadWrite :: (PrimMonad prim, PrimType ty) => Mutable (primState prim) -> ((Offset ty -> prim ty), (Offset ty -> ty -> prim ()))
+primReadWrite addr = (primAddrRead addr, primAddrWrite addr)
+{-# INLINE primReadWrite #-}
