+++ basement/Basement/UArray.hs
@@ -131,8 +131,11 @@ import           Basement.MutableBuilder
 import           Basement.Bindings.Memory (sysHsMemFindByteBa, sysHsMemFindByteAddr)
 import qualified Basement.Compat.ExtList as List
 import qualified Basement.Base16 as Base16
+import qualified Basement.Alg.Native.Prim as PrimBA
 import qualified Basement.Alg.Native.PrimArray as PrimBA
+import qualified Basement.Alg.Foreign.Prim as PrimAddr
 import qualified Basement.Alg.Foreign.PrimArray as PrimAddr
+import qualified Basement.Algorithm as Algorithm
 
 -- | Copy every cells of an existing array to a new array
 copy :: PrimType ty => UArray ty -> UArray ty
@@ -651,13 +654,12 @@ sortBy ford vec = runST $ do
     unsafeFreeze mvec
   where
     !len = length vec
-    !end = 0 `offsetPlusE` len
     !start = offset vec
 
     goNative :: MutableByteArray# (PrimState (ST s)) -> ST s ()
-    goNative mba = PrimBA.inplaceSortBy ford mba start end
+    goNative mba = Algorithm.inplaceSortBy ford start len (PrimBA.primReadWrite mba)
     goAddr :: Ptr ty -> ST s ()
-    goAddr (Ptr addr) = PrimAddr.inplaceSortBy ford addr start end
+    goAddr (Ptr addr) = Algorithm.inplaceSortBy ford start len (PrimAddr.primReadWrite addr)
 {-# SPECIALIZE [3] sortBy :: (Word8 -> Word8 -> Ordering) -> UArray Word8 -> UArray Word8 #-}
 
 filter :: forall ty . PrimType ty => (ty -> Bool) -> UArray ty -> UArray ty
